fix(dashboard version service): List dashboard versions by dashboard UID if populated
If the DashboardUID is populated in the ListDashboardVersionsQuery, use that to get the dashboardID, otherwise use the dashboardID to populated the DashboardUID field in the response.

The GetDashboardVersions API handles this already; it is added as a defensive measure and for consistency with the Get method which will be updated in a future commit.