feat(dashboard versions): populate dashboardUID in the GetDashboardVersions return values
Adds the DashboardUID to the GetDashboardVersionQuery (already in the List query) and populates the dashboardUID in the returned response.

The API gets the DashboardID when querying by UID, but the UID wasn't populated in the returned DashboardVersionDTO. This adds the DashboardUID to GetDashboardVersionQuery, and the DashboardVersions service will use the UID to get the dashboard ID if not set. The DashboardUID is then added to the return values.